requirements.txt
pandas>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
hashlib-additional>=1.0.0; python_version<"3.0"
pytest>=8.0.0
pytest-cov>=5.0.0
psycopg[binary]>=3.2.0
scikit-learn>=1.5.0
joblib>=1.4.0
